requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==2.4.1
billiard==4.2.1
black==24.10.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
cron-descriptor==1.4.5
debugpy==1.8.9
decorator==5.1.1
Django==5.1.3
django-celery-beat==2.7.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-timezone-field==7.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
executing==2.1.0
flake8==7.1.1
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kombu==5.4.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pycountry==24.6.1
pyflakes==3.2.0
Pygments==2.18.0
pytest==8.3.3
pytest-django==4.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
redis==5.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
